+++ ListableUI/Sources/Internal/ItemCell.ContentViewContainer.swift
@@ -9,21 +9,25 @@ import UIKit
 
 
 extension ItemCell {
+    
+    private typealias Side = SwipeActionsView.Side
 
     final class ContentContainerView : UIView {
 
         let contentView : Content.ContentView
-
-        private var swipeConfiguration: SwipeConfiguration?
         
-        var swipeActionsView : SwipeActionsView? {
-            swipeConfiguration?.swipeView
+        private var configurations: [Side: SwipeConfiguration] = [:]
+        
+        private var swipeAccessibilityCustomActions: [Side: [AccessibilitySwipeAction]] = [:] {
+            didSet {
+                updateAccessibilityCustomActions()
+            }
         }
 
         private (set) var swipeState: SwipeActionState = .closed {
             didSet {
                 if oldValue != swipeState {
-                    swipeConfiguration?.swipeView.apply(state: swipeState)
+                    configurations.values.forEach { $0.swipeView.apply(state: swipeState) }
                 }
             }
         }
@@ -44,15 +48,21 @@ extension ItemCell {
         override func layoutSubviews() {
             super.layoutSubviews()
 
-            if let configuration = swipeConfiguration {
-                updateFrames(using: configuration)
-            } else {
+            if configurations.isEmpty {
                 contentView.frame = bounds
+            } else {
+                configurations.values.forEach { updateFrames(using: $0) }
             }
         }
 
         private func updateFrames(using configuration: SwipeConfiguration) {
-
+            
+            let swipeViewSide = configuration.swipeView.side
+            
+            guard swipeState.isRelevantFor(side: swipeViewSide) else {
+                return
+            }
+            
             let xOriginOffset: CGFloat
 
             switch swipeState {
@@ -63,19 +73,33 @@ extension ItemCell {
 
             case .expandActions:
 
-                xOriginOffset = -bounds.width
+                switch swipeViewSide {
+                case .left:
+                    xOriginOffset = bounds.width
+                case .right:
+                    xOriginOffset = -bounds.width
+                }
 
             case .open:
 
-                xOriginOffset = -configuration.swipeView.swipeActionsWidth
+                switch swipeViewSide {
+                case .left:
+                    xOriginOffset = configuration.swipeView.swipeActionsWidth
+                case .right:
+                    xOriginOffset = -configuration.swipeView.swipeActionsWidth
+                }
 
             case .swiping:
 
                 let translation = configuration.panGestureRecognizer.translation(in: self)
-                // No actions exist to the left, so limit overscrolling to the right to 20% of the width.
-                xOriginOffset = min(bounds.width / 5.0, contentView.frame.origin.x + translation.x)
-
                 configuration.panGestureRecognizer.setTranslation(.zero, in: self)
+                
+                switch swipeViewSide {
+                case .left:
+                    xOriginOffset = max(-bounds.width * 0.2, contentView.frame.origin.x + translation.x)
+                case .right:
+                    xOriginOffset = min(bounds.width * 0.2, contentView.frame.origin.x + translation.x)
+                }
 
             case .willPerformFirstActionAutomatically:
 
@@ -85,51 +109,71 @@ extension ItemCell {
 
             contentView.frame = bounds
             contentView.frame.origin.x = xOriginOffset
-            configuration.swipeView.frame = bounds.divided(atDistance: -xOriginOffset, from: .maxXEdge).slice
+                        
+            switch swipeViewSide {
+            case .left:
+                configuration.swipeView.frame = bounds.divided(atDistance: xOriginOffset, from: .minXEdge).slice
+            case .right:
+                configuration.swipeView.frame = bounds.divided(atDistance: -xOriginOffset, from: .maxXEdge).slice
+            }
+        }
+        
+        func isTouchWithinSwipeActionView(touch: UITouch) -> Bool {
+            configurations.values.first { $0.swipeView.contains(touch: touch) } != nil
         }
 
         // MARK: - Swipe Registration
-
-        public func deregisterSwipeIfNeeded() {
-            guard let configuration = swipeConfiguration else { return }
-
-            removeGestureRecognizer(configuration.panGestureRecognizer)
-            configuration.swipeView.removeFromSuperview()
-
-            accessibilityCustomActions = nil
-            swipeConfiguration = nil
-            swipeState = .closed
-
-            setNeedsLayout()
+        
+        func deregisterLeadingSwipeIfNeeded() {
+            deregisterSwipeIfNeeded(for: .left)
         }
 
-        public func registerSwipeActionsIfNeeded(actions: SwipeActionsConfiguration, style: SwipeActionsView.Style, reason: ApplyReason) {
-            if swipeConfiguration == nil {
+        func deregisterTrailingSwipeIfNeeded() {
+            deregisterSwipeIfNeeded(for: .right)
+        }
+        
+        func registerLeadingSwipeActionsIfNeeded(actions: SwipeActionsConfiguration, style: SwipeActionsView.Style, reason: ApplyReason) {
+            registerSwipeActionsIfNeeded(side: .left, actions: actions, style: style, reason: reason)
+        }
+        
+        func registerTrailingSwipeActionsIfNeeded(actions: SwipeActionsConfiguration, style: SwipeActionsView.Style, reason: ApplyReason) {
+            registerSwipeActionsIfNeeded(side: .right, actions: actions, style: style, reason: reason)
+        }
+        
+        private func registerSwipeActionsIfNeeded(
+            side: SwipeActionsView.Side,
+            actions: SwipeActionsConfiguration,
+            style: SwipeActionsView.Style,
+            reason: ApplyReason
+        ) {
+            if configurations[side] == nil {
 
                 let swipeView = SwipeActionsView(
+                    side: side,
                     style: style,
                     didPerformAction: { [weak self] expandActions in
-                        self?.didPerformAction(expandActions: expandActions)
+                        self?.didPerformAction(expandActions: expandActions, side: side)
                     }
                 )
 
                 insertSubview(swipeView, belowSubview: contentView)
                 swipeView.clipsToBounds = true
 
-                let panGestureRecognizer = LeftPanGestureRecognizer(target: self, action: #selector(handlePan))
+                let panGestureRecognizer = DirectionalPanGestureRecognizer(direction: side.gestureDirection, target: self, action: #selector(handlePan))
                 addGestureRecognizer(panGestureRecognizer)
 
-                swipeConfiguration = SwipeConfiguration(
+                configurations[side] = SwipeConfiguration(
                     panGestureRecognizer: panGestureRecognizer,
                     swipeView: swipeView,
                     numberOfActions: actions.actions.count,
-                    performsFirstActionWithFullSwipe: actions.performsFirstActionWithFullSwipe
+                    performsFirstActionWithFullSwipe: actions.performsFirstActionWithFullSwipe,
+                    side: side
                 )
             }
 
-            swipeConfiguration?.numberOfActions = actions.actions.count
-            swipeConfiguration?.swipeView.apply(actions: actions, style: style)
-            configureAccessibilityActions(for: actions.actions)
+            configurations[side]?.numberOfActions = actions.actions.count
+            configurations[side]?.swipeView.apply(actions: actions, style: style)
+            configureAccessibilityActions(actions.actions, for: side)
 
             if reason == .willDisplay {
                 set(state: .closed)
@@ -143,12 +187,25 @@ extension ItemCell {
             if self.listView == nil {
                 self.listView = self.firstSuperview(ofType: ListView.self)
             }
+            
+            guard let configuration = configurations.values.first(
+                where: { $0.panGestureRecognizer == sender }
+            ) else {
+                return
+            }
 
-            guard let configuration = swipeConfiguration else { return }
-
+            let side = configuration.swipeView.side
             let offsetMultiplier = configuration.numberOfActions == 1 ? 0.5 : 0.7
             let performActionOffset = frame.width * CGFloat(offsetMultiplier)
-            let currentSwipeOffset = -contentView.frame.origin.x
+            
+            let currentSwipeOffset: CGFloat
+            switch side {
+            case .left:
+                currentSwipeOffset = contentView.frame.origin.x
+            case .right:
+                currentSwipeOffset = -contentView.frame.origin.x
+            }
+            
             let willPerformAction = currentSwipeOffset > performActionOffset
                 && configuration.performsFirstActionWithFullSwipe
 
@@ -160,43 +217,36 @@ extension ItemCell {
 
             switch sender.state {
             case .began, .changed:
-                let swipeState = SwipeActionState.swiping(willPerformAction: willPerformAction)
+
+                let swipeState = SwipeActionState.swiping(side, willPerformAction: willPerformAction)
                 set(state: swipeState)
 
             case .ended, .cancelled:
 
-                let swipeActionsWidth = configuration.swipeView.swipeActionsWidth
-                let keepOpenOffset = swipeActionsWidth / 2
                 let velocity = sender.velocity(in: self).x
-
+                
+                let isClosing: Bool
+                
+                switch side {
+                case .left:
+                    isClosing = velocity <= 0
+                case .right:
+                    isClosing = velocity >= 0
+                }
+                
                 var swipeState: SwipeActionState
-
-                if velocity < 0 {
-
-                    if willPerformAction {
-                        swipeState = .willPerformFirstActionAutomatically
-                    } else {
-                        swipeState = .open
-                    }
-
-                } else if velocity > 0 {
-
+                
+                if isClosing {
                     swipeState = .closed
-
                 } else {
-
                     if willPerformAction {
-                        swipeState = .willPerformFirstActionAutomatically
-                    } else if currentSwipeOffset > keepOpenOffset {
-                        swipeState = .open
+                        swipeState = .willPerformFirstActionAutomatically(side)
                     } else {
-                        swipeState = .closed
+                        swipeState = .open(side)
                     }
-
                 }
 
                 set(state: swipeState, animated: true)
-
 
             default:
                 set(state: .closed)
@@ -204,10 +254,9 @@ extension ItemCell {
             }
         }
 
-        private func didPerformAction(expandActions: Bool) {
-
+        private func didPerformAction(expandActions: Bool, side: SwipeActionsView.Side) {
             if expandActions {
-                self.set(state: .expandActions, animated: true)
+                self.set(state: .expandActions(side), animated: true)
             } else {
                 self.set(state: .closed, animated: true)
             }
@@ -232,48 +281,100 @@ extension ItemCell {
         }
 
         @objc private func performAccessibilityAction(_ action: AccessibilitySwipeAction) -> Bool {
-            action.action.handler(self.didPerformAction)
+            action.action.handler { _ in
+                self.didPerformAction(expandActions: false, side: action.side)
+            }
             return true
         }
 
-        private func configureAccessibilityActions(for actions: [SwipeAction]) {
-            self.accessibilityCustomActions = actions.map {
-                AccessibilitySwipeAction(action: $0, target: self, selector: #selector(performAccessibilityAction))
+        private func configureAccessibilityActions(_ actions: [SwipeAction], for side: Side) {
+            swipeAccessibilityCustomActions[side] = actions.map {
+                AccessibilitySwipeAction(action: $0, side: side, target: self, selector: #selector(performAccessibilityAction))
             }
+        }
+        
+        private func updateAccessibilityCustomActions() {
+            self.accessibilityCustomActions = swipeAccessibilityCustomActions
+                .values
+                .flatMap { $0 }
+        }
+        
+        private func deregisterSwipeIfNeeded(for side: Side) {
+            guard let configuration = configurations[side] else { return }
+
+            removeGestureRecognizer(configuration.panGestureRecognizer)
+            configuration.swipeView.removeFromSuperview()
+
+            swipeAccessibilityCustomActions[side] = nil
+            configurations[side] = nil
+            swipeState = .closed
+
+            setNeedsLayout()
         }
     }
 
-    struct SwipeConfiguration {
+    private struct SwipeConfiguration {
         let panGestureRecognizer: UIPanGestureRecognizer
         let swipeView: SwipeActionsView
         var numberOfActions: Int
         var performsFirstActionWithFullSwipe: Bool
+        var side: Side
     }
 }
 
 private class AccessibilitySwipeAction: UIAccessibilityCustomAction {
+    typealias Side = SwipeActionsView.Side
+    
     let action: SwipeAction
+    let side: SwipeActionsView.Side
 
-    init(action: SwipeAction, target: Any?, selector: Selector) {
+    init(action: SwipeAction, side: Side, target: Any?, selector: Selector) {
         self.action = action
+        self.side = side
         super.init(name: action.title ?? "", target: target, selector: selector)
     }
 }
 
 /// These states dictate the layout of the swipe actions.
 public enum SwipeActionState: Equatable {
+    public typealias Side = SwipeActionsView.Side
+    
     /// The actions are completely collapsed.
     case closed
 
     /// The actions are open to their natural size.
-    case open
+    case open(Side)
 
     /// The actions are being swiped and the size is affected by the gesture recognizer.
-    case swiping(willPerformAction: Bool)
+    case swiping(Side, willPerformAction: Bool)
 
     /// The actions have been swiped far enough to confirm the first action.
-    case willPerformFirstActionAutomatically
+    case willPerformFirstActionAutomatically(Side)
 
     /// The actions have been asked to completely expand (typically because the item is being deleted).
-    case expandActions
+    case expandActions(Side)
+    
+    func isRelevantFor(side: Side) -> Bool {
+        switch self {
+        case .closed:
+            return true
+        case .open(let stateSide),
+                .swiping(let stateSide, _),
+                .willPerformFirstActionAutomatically(let stateSide),
+                .expandActions(let stateSide):
+            return stateSide == side
+        }
+    }
+}
+
+private extension SwipeActionsView.Side {
+    
+    var gestureDirection: DirectionalPanGestureRecognizer.Direction {
+        switch self {
+        case .left:
+            return .leftToRight
+        case .right:
+            return .rightToLeft
+        }
+    }
 }