+++ ListableUI/Sources/Internal/SwipeActionsView.swift
@@ -10,6 +10,11 @@ import UIKit
 private let haptics = UIImpactFeedbackGenerator(style: .light)
 
 public final class SwipeActionsView: UIView {
+    
+    public enum Side: Equatable {
+        case left
+        case right
+    }
 
     public struct Style: Equatable {
         public enum Shape: Equatable {
@@ -97,6 +102,9 @@ public final class SwipeActionsView: UIView {
 
     private var state: SwipeActionState = .closed
     
+    /// The side this swipe actions view will originate from when presented.
+    let side: Side
+
     private var availableButtonWidth: CGFloat {
         guard let superview else {
             return .greatestFiniteMagnitude
@@ -106,9 +114,11 @@ public final class SwipeActionsView: UIView {
     }
 
     public init(
+        side: Side,
         style: Style,
         didPerformAction: @escaping SwipeAction.CompletionHandler
     ) {
+        self.side = side
         self.style = style
         self.didPerformAction = didPerformAction
         super.init(frame: .zero)
@@ -131,17 +141,25 @@ public final class SwipeActionsView: UIView {
         container.frame.size.height = max(0, bounds.size.height - insets.top - insets.bottom)
         
         container.layer.cornerRadius = style.containerCornerRadius
-
+        
+        let buttons: [DefaultSwipeActionButton]
+        switch side {
+        case .left:
+            buttons = actionButtons.reversed()
+        case .right:
+            buttons = actionButtons
+        }
+        
         // Calculates the x origin for each button based on the width of each button before it
         // and the percent that the actions are slid open for the overlapping parallax effect
         func xOriginForButton(at index: Int) -> CGFloat {
-            let previousButtons = Array(actionButtons[0..<index])
+            let previousButtons = Array(buttons[0..<index])
             let position = width(ofButtons: previousButtons)
             let percentOpen = bounds.width / swipeActionsWidth
             return percentOpen * position
         }
 
-        for (index, button) in actionButtons.enumerated() {
+        for (index, button) in buttons.enumerated() {
             button.frame.size.width = width(ofButtons: [button])
             button.frame.size.height = container.bounds.height
 
@@ -152,26 +170,52 @@ public final class SwipeActionsView: UIView {
             wrapperView.frame = button.frame
             wrapperView.frame.origin.x = xOriginForButton(at: index) + CGFloat(index) * style.interActionSpacing
             wrapperView.frame.size.width = max(0, xOriginForButton(at: index + 1) - xOriginForButton(at: index))
-
-            // If there's only one action, the button stays right-aligned while the container stretches.
-            // For multiple actions, they stay left-aligned.
-            if wrapperView.frame.width > button.frame.width && actionButtons.count == 1 {
-                button.frame.origin.x = wrapperView.frame.width - button.frame.width
-            } else {
+            
+            func alignLeftEdge() {
                 button.frame.origin.x = 0
+            }
+            
+            func alignRightEdge() {
+                button.frame.origin.x = wrapperView.frame.width - button.frame.width
+            }
+            
+            // If there's only one action, the button stays aligned with the outer edge
+            // while the container stretches.
+            // For multiple actions, they stay aligned to the inner edge.
+            if wrapperView.frame.width > button.frame.width && buttons.count == 1 {
+                switch side {
+                case .left:
+                    alignLeftEdge()
+                case .right:
+                    alignRightEdge()
+                }
+            } else {
+                switch side {
+                case .left:
+                    alignRightEdge()
+                case .right:
+                    alignLeftEdge()
+                }
             }
         }
 
         // Adjust the last button container view to fill the safe area space
-        if let lastButtonContainer = actionButtons.last?.superview {
+        if let lastButtonContainer = buttons.last?.superview {
             lastButtonContainer.frame.size.width = max(0, container.bounds.width - lastButtonContainer.frame.origin.x)
         }
 
         // If the last action will be automatically performed or the state is set to expand the actions
         // for performing the last action, have the last action fill the available space.
-        if state == .swiping(willPerformAction: true) || state == .expandActions {
+        if state == .swiping(side, willPerformAction: true) || state == .expandActions(side) {
             actionButtons.last?.superview?.frame = container.bounds
-            actionButtons.last?.frame.origin.x = 0
+            
+            switch side {
+            case .left:
+                actionButtons.last?.frame.origin.x = container.bounds.maxX - (actionButtons.last?.frame.width ?? 0)
+            case .right:
+                actionButtons.last?.frame.origin.x = 0
+            }
+            
         }
     }
 
@@ -232,13 +276,18 @@ public final class SwipeActionsView: UIView {
         calculatedNaturalWidth = width(ofButtons: actionButtons) + style.containerInsets.left + style.containerInsets.right
     }
 
-    public func apply(state: SwipeActionState) {
+    public func apply(state newState: SwipeActionState) {
+        let priorState = state
+        state = newState
+        
+        guard newState.isRelevantFor(side: side) else {
+            return
+        }
+        
         haptics.prepare()
 
-        switch (state, self.state) {
-        case (.swiping, .swiping) where state != self.state:
-
-            self.state = state
+        switch (newState, priorState) {
+        case (.swiping, .swiping) where newState != priorState:
 
             haptics.impactOccurred()
 
@@ -249,13 +298,9 @@ public final class SwipeActionsView: UIView {
 
         case (.willPerformFirstActionAutomatically, _):
 
-            firstAction.flatMap { action in
-                action.handler(didPerformAction)
-            }
+            firstAction?.handler(didPerformAction)
 
         default:
-
-            self.state = state
 
             setNeedsLayout()